Refactors project into demo project and library
