+++ .idea/modules.xml
@@ -4,6 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/TouchImageView.iml" filepath="$PROJECT_DIR$/TouchImageView.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/touchview/touchview.iml" filepath="$PROJECT_DIR$/touchview/touchview.iml" />
     </modules>
   </component>
 </project>
