+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "21.1.2"
 
     defaultConfig {
-        applicationId "com.example.touch"
+        applicationId "com.ortiz.touchdemo"
         minSdkVersion 8
         targetSdkVersion 22
     }
@@ -20,4 +20,5 @@ android {
 
 dependencies {
     compile 'com.android.support:support-v4:22.0.0'
+    compile project(':touchview')
 }
