+++ app/src/main/java/com/ortiz/touchdemo/MainActivity.java
@@ -1,12 +1,12 @@
-package com.ortiz.touch;
-
-import com.example.touch.R;
+package com.ortiz.touchdemo;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.view.View.OnClickListener;
+
+import com.example.touch.R;
 
 
 public class MainActivity extends Activity {
