+++ app/src/main/java/com/ortiz/touchdemo/SwitchImageExampleActivity.java
@@ -1,11 +1,12 @@
-package com.ortiz.touch;
-
-import com.example.touch.R;
+package com.ortiz.touchdemo;
 
 import android.app.Activity;
 import android.os.Bundle;
 import android.view.View;
 import android.view.View.OnClickListener;
+
+import com.example.touch.R;
+import com.ortiz.touchview.TouchImageView;
 
 public class SwitchImageExampleActivity extends Activity {
