+++ app/src/main/java/com/ortiz/touchdemo/SwitchScaleTypeExampleActivity.java
@@ -1,13 +1,14 @@
-package com.ortiz.touch;
-
-import com.example.touch.R;
+package com.ortiz.touchdemo;
 
 import android.app.Activity;
 import android.os.Bundle;
 import android.view.View;
 import android.view.View.OnClickListener;
-import android.widget.Toast;
 import android.widget.ImageView.ScaleType;
+import android.widget.Toast;
+
+import com.example.touch.R;
+import com.ortiz.touchview.TouchImageView;
 
 public class SwitchScaleTypeExampleActivity extends Activity {
