+++ app/src/main/res/layout/activity_mirroring_example.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical" >
     
-    <com.ortiz.touch.TouchImageView 
+    <com.ortiz.touchview.TouchImageView
         android:id="@+id/topImage"
         android:src="@drawable/nature_1"
         android:layout_width="wrap_content"
@@ -16,7 +16,7 @@
         android:layout_width="match_parent"
         android:layout_height="1dp" />
     
-    <com.ortiz.touch.TouchImageView 
+    <com.ortiz.touchview.TouchImageView
         android:id="@+id/bottomImage"
         android:src="@drawable/nature_2"
         android:layout_width="wrap_content"
