+++ app/src/main/res/layout/activity_switch_scaletype_example.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent" >
     
-    <com.ortiz.touch.TouchImageView 
+    <com.ortiz.touchview.TouchImageView
         android:id="@+id/img"
         android:scaleType="center"
         android:src="@drawable/nature_1"
