+++ app/src/main/res/layout/activity_viewpager_example.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent" >
 
     <!-- ExtendedViewPager is a custom view and must be referred to by its full package name in XML -->
-	<com.ortiz.touch.ExtendedViewPager 
+	<com.ortiz.touchdemo.ExtendedViewPager
 	    android:id="@+id/view_pager"
 	    android:layout_width="match_parent"
 	    android:layout_height="match_parent" />
