+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app'
+include ':app', ':touchview'
